Loop induction variable should not be int16_t, or we crash on large grids!

+++ runtime.c
@@ -6,7 +6,7 @@ int16_t cell(int16_t *oldgrid, int16_t *newgrid, int16_t width, int16_t height, 
 void automaton(int16_t *oldgrid, int16_t *newgrid, int16_t width, int16_t
     height) {
   int16_t g[10] = {0};
-  int16_t i=0;
+  int i=0;
   for (int16_t x=0 ; x<width ; x++) {
     for (int16_t y=0 ; y<height ; y++,i++) {
       newgrid[i] = cell(oldgrid, newgrid, width, height, x, y, oldgrid[i], g);
